Corrige problèmes identifiés par eslint
