+++ client/src/pages/Teacher/ManageRoom/RoomContext.tsx
@@ -1,16 +1,8 @@
-import { createContext, useContext, useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import ApiService from '../../../services/ApiService';
 import { RoomType } from 'src/Types/RoomType';
 import React from "react";
-
-type RoomContextType = {
-  rooms: RoomType[];
-  selectedRoom: RoomType | null;
-  selectRoom: (roomId: string) => void;
-  createRoom: (title: string) => Promise<void>;
-};
-
-const RoomContext = createContext<RoomContextType | undefined>(undefined);
+import { RoomContext } from './useRooms';
 
 export const RoomProvider = ({ children }: { children: React.ReactNode }) => {
   const [rooms, setRooms] = useState<RoomType[]>([]);
@@ -64,10 +56,4 @@ export const RoomProvider = ({ children }: { children: React.ReactNode }) => {
       {children}
     </RoomContext.Provider>
   );
-};
-
-export const useRooms = () => {
-  const context = useContext(RoomContext);
-  if (!context) throw new Error('useRooms must be used within a RoomProvider');
-  return context;
 };
